chore: move IAppAttachmentValueProvider to AeFinder.Sdk
