+++ test/AeFinder.App.Tests/Attachments/AppAttachmentValueProviderTests.cs
@@ -1,7 +1,9 @@
 using System.Threading.Tasks;
 using AeFinder.AmazonCloud;
+using AeFinder.Apps;
 using AeFinder.Grains;
 using AeFinder.Grains.Grain.Subscriptions;
+using AeFinder.Sdk.Attachments;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 using Orleans;
@@ -15,16 +17,18 @@ public sealed class AppAttachmentValueProviderTests : AeFinderAppTestBase
     private readonly IAwsS3ClientService _awsS3ClientService;
     private readonly IClusterClient _clusterClient;
     private readonly IAppInfoProvider _appInfoProvider;
+    private readonly IAppAttachmentService _appAttachmentService;
 
     public AppAttachmentValueProviderTests()
     {
         _awsS3ClientService = GetRequiredService<IAwsS3ClientService>();
         _clusterClient = GetRequiredService<IClusterClient>();
         _appInfoProvider = GetRequiredService<IAppInfoProvider>();
+        _appAttachmentService = GetRequiredService<IAppAttachmentService>();
     }
 
     [Fact]
-    public async Task InitTest()
+    public async Task InitAppAttachmentValuesTest()
     {
         var fileName = "testFileName.json";
         var key = "TestKey";
@@ -33,12 +37,15 @@ public sealed class AppAttachmentValueProviderTests : AeFinderAppTestBase
         var providers = ServiceProvider.GetServices<IAppAttachmentValueProvider>();
         foreach (var provider in providers)
         {
-            await provider.InitValueAsync();
+            var content =
+                await _appAttachmentService.GetAppAttachmentContentAsync(_appInfoProvider.AppId,
+                    _appInfoProvider.Version, provider.Key);
+            provider.InitValue(content);
         }
 
         var appAttachmentValueProvider = GetRequiredService<IAppAttachmentValueProvider<TestInfo>>();
-        var content = await _awsS3ClientService.GetJsonFileContentAsync(_appInfoProvider.AppId, fileName);
-        var value = JsonConvert.DeserializeObject<TestInfo>(content);
+        var fileContent = await _awsS3ClientService.GetJsonFileContentAsync(_appInfoProvider.AppId, fileName);
+        var value = JsonConvert.DeserializeObject<TestInfo>(fileContent);
         appAttachmentValueProvider.GetValue().Info.ShouldBe(value!.Info);
     }
 }
